+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## 20.11.11
+  * Deprecating "ADVERTISING_ID" as device ID generation strategy. SDK will fall back to 'OPEN_UDID'.
   * Orientation tracking is not enabled by default. Can be turned off during init.
   * Device ID can now be changed when no consent is given
   * Push notification now display/use the sent badge number. It's visualization depends on the launcher.